fix: broken tests + code refactor
